add context global
