+++ packages/container-registry-v5/test/helpers.js
@@ -16,22 +16,3 @@ if (process.env.ENABLE_NET_CONNECT === 'true') {
 
 const chai = require('chai')
 chai.use(require('chai-as-promised'))
-
-const expect = require('chai').expect
-
-function assertExit(code, gen) {
-  let actualError
-  return gen.catch(function (error) {
-    expect(error).to.be.an.instanceof(cli.exit.ErrorExit)
-    actualError = error
-  }).then(function () {
-    expect(actualError).to.not.equal(undefined, 'Expected Error to be thrown')
-    expect(actualError.code).to.be.an('number', 'Expected error.exit(i) to be called with a number')
-    expect(actualError.code).to.equal(code)
-    return actualError
-  })
-}
-
-module.exports = {
-  assertExit,
-}
